gate fuzz testing behind build tags requiring Go 1.18 or higher
